+++ projects/wickeyappstore/src/lib/api-connection.service.ts
@@ -536,7 +536,7 @@ export class ApiConnectionService {
    * @param _params string: user_id, json: was_data where was_data is format {key:value,...}
    * @returns Returns rank of user.
    */
-  setHighscore(_params: { user_id: string, highscore: number }): Observable<{ rank?: number }> {
+  setHighscore(_params: { user_id: string, highscore: number }): Observable<{ status: number, rank?: number }> {
     this.handleHeaders();
     // NOTE: Use share to avoid duplicate calls
     return this.http.post(this.highscoreUrl, _params, { headers: this.apiHeaders, withCredentials: true }).pipe(